+++ src/ideal/ideal.jl
@@ -1240,7 +1240,7 @@ function vdim(I::sideal{S}) where {T <: Nemo.FieldElem,
                                    S <: Union{spoly{T}, spluralg{T}}}
    I.isGB || error("Not a Groebner basis")
    R = base_ring(I)
-   GC.@preserve I R return Int(libSingular.id_vdim(I.ptr, R.ptr))
+   GC.@preserve I R return Int64(libSingular.id_vdim(I.ptr, R.ptr))
 end
 
 @doc raw"""
